requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow
django-filter
drf-yasg